[loader] Fix a crash on actual DVD-R9 multi-game discs.
Similar to the ISO fix, but for RealDI.

FIXME: When testing my old DVD-R9, every game kept complaining that the
disc cover was open. Removing the disc and reinserting it then resulted
in the game claiming it was the wrong disc.

+++ loader/include/dip.h
@@ -66,7 +66,7 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 #define		DIP_CMD_NORMAL	0xA8
 #define		DIP_CMD_DVDR	0xD0
 
-void ReadRealDisc(u8 *Buffer, u32 Offset, u32 Length, u32 Command);
+void ReadRealDisc(u8 *Buffer, u64 Offset, u32 Length, u32 Command);
 void DVDStartCache(void);
 
 #endif
