+++ loader/source/dip.c
@@ -33,16 +33,16 @@ void DVDStartCache(void)
 //Use same buffer as kernel, uncached though, to see howto cache see DI.c
 static u8 *const DISC_DRIVE_BUFFER = (u8*)0x92000800;
 static const u32 DISC_DRIVE_BUFFER_LENGTH = 0x7FF000;
-void ReadRealDisc(u8 *Buffer, u32 Offset, u32 Length, u32 Command)
+void ReadRealDisc(u8 *Buffer, u64 Offset, u32 Length, u32 Command)
 {
 	u32 ReadDiff = 0;
 
 	u32 TmpLen = Length;
-	u32 TmpOffset = Offset;
+	u64 TmpOffset = Offset;
 	if(Command == DIP_CMD_DVDR)
 	{
 		TmpOffset = ALIGN_BACKWARD(Offset, 0x800);
-		ReadDiff = Offset - TmpOffset;
+		ReadDiff = (u32)(Offset - TmpOffset);
 		TmpLen = ALIGN_FORWARD(TmpLen + ReadDiff, 0x800);
 	}
 
@@ -50,7 +50,7 @@ void ReadRealDisc(u8 *Buffer, u32 Offset, u32 Length, u32 Command)
 
 	//Actually read
 	write32(DIP_CMD_0, Command << 24);
-	write32(DIP_CMD_1, Command == DIP_CMD_DVDR ? TmpOffset >> 11 : TmpOffset >> 2);
+	write32(DIP_CMD_1, Command == DIP_CMD_DVDR ? (u32)(TmpOffset >> 11) : (u32)(TmpOffset >> 2));
 	write32(DIP_CMD_2, Command == DIP_CMD_DVDR ? TmpLen >> 11 : TmpLen);
 
 	DCInvalidateRange(DISC_DRIVE_BUFFER, TmpLen);
